requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==1.11
djangorestframework==3.9.1
idna==2.8
Naked==0.1.31
pycrypto==2.6.1
pycryptodome==3.7.3
pytz==2018.9
PyYAML==3.13
requests==2.21.0
shellescape==3.4.1
urllib3==1.24.1
